+++ src/Ampersand/Exception/InterfaceNotDefined.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ampersand\Exception;
+
+use Ampersand\Exception\AmpersandException;
+
+class InterfaceNotDefined extends AmpersandException
+{
+}