+++ src/Ampersand/Interfacing/Resource.php
@@ -72,6 +72,11 @@ class Resource extends Atom implements ArrayAccess
         return (string) parent::jsonSerialize();
     }
 
+    public function getLabel(): string
+    {
+        return implode('', $this->ifc->getViewData($this));
+    }
+
     /**
      * Return interface for this resource
      *
@@ -349,6 +354,6 @@ class Resource extends Atom implements ArrayAccess
 
     public function remove(): void
     {
-        $this->parentList->remove($this);
+        $this->parentList->remove($this->id);
     }
 }