+++ src/Ampersand/Rule/ExecEngine.php
@@ -113,6 +113,11 @@ class ExecEngine extends RuleEngine
         return $this->ampersandApp;
     }
 
+    public function getLogger(): LoggerInterface
+    {
+        return $this->logger;
+    }
+
     /**
      * Logs with an arbitrary level.
      *
@@ -194,13 +199,12 @@ class ExecEngine extends RuleEngine
             
             // Fix violations
             $total = count($violations);
-            $this->info("{++ ExecEngine fixing {$total} violations for rule '{$rule}'");
+            $this->info("ExecEngine fixing {$total} violations for rule '{$rule}'");
             foreach ($violations as $key => $violation) {
                 /** @var \Ampersand\Rule\Violation $violation */
                 $num = $key + 1;
-                $this->info("{+++ Fixing violation {$num}/{$total}: ({$violation})");
+                $this->info("Fixing violation {$num}/{$total}: ({$violation})");
                 $this->fixViolation($violation);
-                $this->info("+++}");
                 
                 // Abort loop when exec engine is terminated
                 if ($this->isTerminated) {
@@ -209,7 +213,7 @@ class ExecEngine extends RuleEngine
                 }
             }
             $rulesFixed[] = $rule;
-            $this->info("++} ExecEngine fixed {$total} violations for rule '{$rule}'");
+            $this->notice("ExecEngine fixed {$total} violations for rule '{$rule}'");
         }
 
         return $rulesFixed;