+++ changelog.md
@@ -2,6 +2,21 @@
 
 ## Unreleased changes
 
+## v1.4.0 (3 january 2020)
+* Bugfix issue with API interfaces shown in UI to solve signal violations. Caused by wrongly placed parentheses.
+* Add sort values for all BOX templates that start with the char 'S' (for SORT). Instead of only for the SCOLS, SHCOLS and SPCOLS templates.
+* [Issue 1005](https://github.com/AmpersandTarski/Ampersand/issues/1005) Bugfix deadlock due to un-defined interfaces
+* [Issue 426](https://github.com/AmpersandTarski/Ampersand/issues/426) Add support for optimized VIEW expression queries using injection of _SRCATOM
+* Add Dockerfile to build Ampersand prototype framework image that can be used by containerized prototype apps
+* Improve configuration of prototype for different environments (dev/prod/...). See [readme](./config/README.md)
+* Simplify logging. Log to stdout and stderr to work with containerized prototype apps
+* Add logs for add/rm/del atom and add/del links
+* Add CI using Github Actions to build docker image and perform php static analysis using phan
+* [Issue 940](https://github.com/AmpersandTarski/Ampersand/issues/940) Remove company logos from footer
+* [Issue 951](https://github.com/AmpersandTarski/Ampersand/issues/951) Refresh page after role (de)select
+* [Issue 983](https://github.com/AmpersandTarski/Ampersand/issues/983) Increase default timeout for installing the application to 5 min
+* [Issue 1016](https://github.com/AmpersandTarski/Ampersand/issues/1016) Fix for invariant violation in metapopulation that will be resolved by initial population. Installing application is now in a single transaction
+
 ## v1.3.0 (15 july 2019)
 * Bugfix error message in case of network/connection error
 * Many bugfixes (see commit history)